[translation] Translate the wait builtin
